requirements.txt
# Core dependencies for HRRR ASNOW Zarr processing
zarr>=2.16.0
xarray>=2023.1.0
cfgrib>=0.9.10
eccodes>=1.6.0
numpy>=1.24.0
pyproj>=3.5.0
dask>=2023.1.0
requests>=2.28.0
pyyaml>=6.0
python-dateutil>=2.8.2

# Validation and visualization
matplotlib>=3.7.0
pandas>=2.0.0

# Optional but recommended
s3fs>=2023.1.0
boto3>=1.26.0
